+++ src/index.ts
@@ -2,3 +2,6 @@ import "../style.css";
 
 export { Parallax } from "./Parallax";
 export type { ParallaxProps, ParallaxAxis } from "./Parallax";
+
+export { FadeOnView } from "./FadeOnView";
+export type { FadeOnViewProps } from "./FadeOnView";